testapp/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
django-seed==0.3.1
Django==4.1
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==20.0.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pip==23.3.1
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml.clib==0.2.8
ruamel.yaml==0.18.5
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.41.3
